Create prize manager class
