+++ paper/src/main/java/com/badbones69/crazycrates/paper/cratetypes/CSGO.java
@@ -121,7 +121,7 @@ public class CSGO implements Listener {
                         crazyManager.endCrate(player);
                         Prize prize = crate.getPrize(inv.getItem(13));
 
-                        plugin.getCrazyHandler().checkPrize(prize, player, crate);
+                        plugin.getCrazyHandler().getPrizeManager().checkPrize(prize, player, crate);
 
                         crazyManager.removePlayerFromOpeningList(player);
                         cancel();
